==> Master_Thesis/Step 0. Convert .pkl to .csv file.ipynb <==
import os
import pandas as pd

directory = os.getcwd()
path = directory + '/Data/'

# Reading .pkl files

Filing_final=pd.read_pickle(path + 'Filing_final.pkl', compression='bz2')
Hold_final=pd.read_pickle(path + 'Hold_final.pkl', compression='bz2')
Filings_df=pd.read_pickle(path + 'Filings_df.pkl', compression='bz2')
Holdings=pd.read_pickle(path + 'Holdings.pkl', compression='bz2')
Holdings_df=pd.read_pickle(path + 'Holdings_df.pkl', compression='bz2')
cik_v3=pd.read_pickle(path+'cik_v3.pkl', compression='bz2')
Relationships=pd.read_pickle(path+'Relationships.pkl', compression='bz2')
Relationships_wo_options=pd.read_pickle(path+'Relationships_wo_options.pkl', compression='bz2')

# Converting to .csv files

Filing_final.to_csv(path + 'Filing_final.csv', index=False)
Hold_final.to_csv(path + 'Hold_final.csv', index=False)
Filings_df.to_csv(path + 'Filings_df.csv', index=False)
Holdings.to_csv(path + 'Holdings.csv', index=False)
Holdings_df.to_csv(path + 'Holdings_df.csv', index=False)
cik_v3.to_csv(path + 'cik_v3.csv', index=False)
Relationships.to_csv(path + 'Relationships.csv', index=False)
Relationships_wo_options.to_csv(path + 'Relationships_wo_options.csv', index=False)
