==> Master_Thesis/Step 2.2 Data Preprocessing(Fuzzy matcher).ipynb <==
import import_ipynb
from ipynb.fs.full.Functions_sec import *
import fuzzymatcher as fm 
import pandas as pd
import os

directory = os.getcwd()
path = directory + '/Data/'

# Reading files

cik_header=["nameOfIssuer","CIK"]
cik_v3=pd.read_csv(path + 'cik_v3.csv',low_memory=False)
cik_final=pd.read_csv(path + 'cik_final.csv',names=cik_header,header=0,sep=";")

# Joining using fuzzymatcher

cik_v4=cik_v3[cik_v3["CIK"].isna()]
for_matching=cik_v4["nameOfIssuer_x"].unique()
for_matching=pd.DataFrame(for_matching, columns=(["nameOfIssuer_x"]))

# Columns to match on from df_left
left_on = ["nameOfIssuer_x"]

# Columns to match on from df_right
right_on = ["nameOfIssuer"]

df=split_dataframe(for_matching,1000)

#matched_results
for x in range(len(df)):
    print(x)
    a = fm.fuzzy_left_join(df[x], cik_final,left_on,right_on)
    if x==0:
        matched_results_final = a 
    else:
        matched_results_final=pd.concat((matched_results_final,a), axis=0,ignore_index=True)

matched_results_final.to_csv(path + 'Matched.csv')
