==> Master_Thesis/Step 2.3 Data Preprocessing.ipynb <==
import pandas as pd
import os

directory = os.getcwd()
path = directory + '/Data/'

# Reading files

Matched=pd.read_csv(path + 'Matched_v4.csv', sep=';')
cik_v3=pd.read_csv(path + 'cik_v3.csv',low_memory=False)
Holdings_df=pd.read_csv(path + 'Holdings_df.csv')

cik_v4=cik_v3[cik_v3["CIK"].isna()]
for_matching=cik_v4["nameOfIssuer_x"].unique()
for_matching=pd.DataFrame(for_matching, columns=(["nameOfIssuer_x"]))

# Joining fuzzy match records with Holdings data

#a=pd.merge(Matched, df[0], how='right', on='nameOfIssuer_x')
cik_v4=cik_v4.drop(["CIK","nameOfIssuer_unchanged","nameOfIssuer_y","nameOfIssuer_1"],axis=1)

merge_=pd.merge(cik_v4,Matched, how='left', on='nameOfIssuer_x')

merge_5=merge_[~merge_["CIK"].isna()][["cusip","cik_Filing","titleOfClass","putCall","CIK","investmentDiscretion","value","nameOfIssuer_x","nameOfIssuer"]]
merge_5.rename(columns = {'nameOfIssuer':'nameOfIssuer_unchanged','nameOfIssuer_x':'nameOfIssuer'}, inplace = True)
#merge_5.rename(columns = {'nameOfIssuer':'nameOfIssuer_x'}, inplace = True)

Holdings=pd.concat((Holdings_df,merge_5),axis=0, ignore_index=True).reset_index(drop=True)

# Saving Holdings file

Holdings.to_csv(path + 'Holdings.csv', index=False)
