==> Master_Thesis/Step 3.1 Nodes & Relationships creation.ipynb <==
#pip install sqlite-fts4
#pip install import-ipynb

from ipynb.fs.full.Functions_sec import *
import import_ipynb
import fuzzymatcher as fm 
import pandas as pd
import os
pd.set_option('display.max_rows', 1200)

directory = os.getcwd()
path = directory + '/Data/'

# Reading files

Holdings=pd.read_csv(path + 'Holdings.csv')
Matched=pd.read_csv(path + 'Matched_v4.csv', sep=';')
Filings=pd.read_csv(path + 'Filings_df.csv')

# Nodes and Relationships Extraction

Nodes=Nodes_(Holdings,Filings)
duplicates_company_name=Nodes[1]
Nodes=Nodes[0]

Relationships=Relationships_(Holdings)

Relationships[(Relationships['value']==0)]

Relationships=Relationships[~(Relationships['value']==0)]

Relationships=Relationships.groupby(by=['cik_Filing','CIK','titleOfClass','putCall','investmentDiscretion'],dropna=False).sum()
Relationships=Relationships.reset_index()

Relationships.to_csv(path + 'Relationships.csv', index=None)
Nodes.to_csv( path + "Nodes.csv" ,index=None)

## Validation of nodes

Relationships[~Relationships["cik_Filing"].isin (Nodes["cik"])]["cik_Filing"].unique()

Relationships[~Relationships["CIK"].isin (Nodes["cik"])]["CIK"].unique()

len(Relationships[~Relationships["CIK"].isin (Nodes["cik"])]["CIK"].unique())
# add these 3 

len(Relationships[~Relationships["cik_Filing"].isin (Nodes["cik"])]["cik_Filing"].unique())
# add these 6

## Notional value of derivative options

#different relationships for call and put, means different labels
Relationships_options=Relationships[~Relationships["putCall"].isna()]
#concat and then unique
a=pd.DataFrame(Relationships_options["CIK"].unique())
b=pd.DataFrame(Relationships_options["cik_Filing"].unique())
h=pd.concat((a,b),axis=0)
h.columns=["cik"]

h=pd.DataFrame(h["cik"].unique())
h.columns=["cik"]

Nodes_options=pd.merge(h,Nodes, how="left")

Relationships_options.to_csv(path + 'Relationships_options.csv')
Nodes_options.to_csv(path + 'Nodes_options.csv')

## Without derivative options

Relationships_wo_options=Relationships[Relationships["putCall"].isna()]
#concat and then unique
a=pd.DataFrame(Relationships_wo_options["CIK"].unique())
b=pd.DataFrame(Relationships_wo_options["cik_Filing"].unique())
h=pd.concat((a,b),axis=0)
h.columns=["cik"]

h=pd.DataFrame(h["cik"].unique())
h.columns=["cik"]

Nodes_wo_options=pd.merge(h,Nodes, how="left")

Relationships_wo_options.to_csv(path + 'Relationships_wo_options.csv')
Nodes_wo_options.to_csv( path +'Nodes_wo_options.csv')

## Relationships with self loops

len(Relationships[(Relationships["cik_Filing"])==(Relationships["CIK"])]) #check if self loops are between banks only

Self_loops=Relationships[(Relationships["cik_Filing"])==(Relationships["CIK"])]
self_loops_=pd.merge(Self_loops, Nodes, left_on="CIK", right_on="cik")

self_loops_

Filings[Filings['cik']==804328]

Holdings[Holdings['cik_Filing']==804328]
