==> Master_Thesis/Step 4.2 Centrality_measures_graphoptions.ipynb <==
from neo4j import GraphDatabase
import pandas as pd
from tabulate import tabulate
import os
import matplotlib
matplotlib.use('TkAgg')
import matplotlib.pyplot as plt

directory = os.getcwd()
path = directory + '/Data/'

## Connecting to neo4j

To connect to neo4j database, lauch your database in the neo4j application. Use your password in the following statement

driver=GraphDatabase.driver("bolt://localhost:7687",auth=("neo4j","<Password>"))

driver=GraphDatabase.driver("bolt://localhost:7687",auth=("neo4j",""))
session = driver.session(database="graphoptions")

result={"label":[],"count":[]}
with driver.session(database="graphoptions") as session:
    labels=[row["label"] for row in session.run("CALL db.labels()")]
    for label in labels:
        query=f"MATCH (:`{label}`) RETURN count(*) as count"
        count=session.run(query).single()["count"]
        result["label"].append(label)
        result["count"].append(count)
df=pd.DataFrame(data=result)
print(tabulate(df.sort_values("count"), headers='keys',tablefmt='psql',showindex=False))

## Degree centrality: Counts outgoing relationships
Detects the number of direct connections a node has


pd.set_option('display.max_rows', 500)
fig=plt.figure()
query="""CALL gds.degree.stream({
nodeProjection: 'company',
relationshipProjection: {Holds: {type: 'Holds'}}})
YIELD nodeId, score
RETURN gds.util.asNode(nodeId).id AS cik,gds.util.asNode(nodeId).company AS name, score as Degree_natural
ORDER BY Degree_natural DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_natural=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_natural=Degree_natural[['cik','name','Degree_natural']]
Degree_natural

query="""CALL gds.degree.stream({
nodeProjection: 'company',
relationshipProjection: {Holds: {type: 'Holds'}}, orientation: 'REVERSE'  })
YIELD nodeId, score
RETURN gds.util.asNode(nodeId).id AS cik,   score as Degree_reverse
ORDER BY Degree_reverse DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_reverse=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_reverse=Degree_reverse[['cik','Degree_reverse']]
Degree_reverse

query='''match ()-[r:Holds]->()
where r.value = toString(r.value)
set r.value = toInteger(r.value)
return count(r)'''
with driver.session(database="graphoptions") as session:
    session.run(query)

query="""CALL gds.degree.stream(({
  nodeProjection: 'company',
  relationshipProjection: {
    relType: {
      type: 'Holds',
      orientation: 'NATURAL',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value'
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik,  score as Degree_natural_wt
RETURN cik,  Degree_natural_wt
ORDER BY Degree_natural_wt DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_natural_wt=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_natural_wt=Degree_natural_wt[['cik','Degree_natural_wt']]
Degree_natural_wt

query="""CALL gds.degree.stream(({
  nodeProjection: 'company',
  relationshipProjection: {
    relType: {
      type: 'Holds',
      orientation: 'REVERSE',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value'
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik,   score as Degree_reverse_wt
RETURN cik,  Degree_reverse_wt
ORDER BY Degree_reverse_wt DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_reverse_wt=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_reverse_wt=Degree_reverse_wt[['cik','Degree_reverse_wt']]
Degree_reverse_wt

## Betweenness(unweighted)
A way of detecting the amount of influence a node has over the flow of information in a graph


query='''CALL gds.betweenness.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'NATURAL',
      properties: {}
    }
  }
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik,  score as Betweeness
RETURN cik,Betweeness
ORDER BY Betweeness DESC
'''

with driver.session(database="graphoptions") as session:
    Betweeness=pd.DataFrame([dict(record) for record in session.run(query)])
    Betweeness=Betweeness[['cik','Betweeness']]
Betweeness

query='''CALL gds.betweenness.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'REVERSE',
      properties: {}
    }
  }
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik,  score as Betweeness_reverse
RETURN cik, Betweeness_reverse
ORDER BY Betweeness_reverse DESC
'''

with driver.session(database="graphoptions") as session:
    Betweeness_reverse=pd.DataFrame([dict(record) for record in session.run(query)])
    Betweeness_reverse=Betweeness_reverse[['cik','Betweeness_reverse']]
Betweeness_reverse

## Approx betweeness
calculates shortest paths between a subset of nodes, unlike Betweenness which considers all pairs of nodes

query='''CALL gds.betweenness.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'NATURAL',
      properties: {}
    }
  },
  samplingSize: 100
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik,  score 
RETURN cik,  score as Ap_Betweeness_natural
ORDER BY Ap_Betweeness_natural DESC
'''

with driver.session(database="graphoptions") as session:
    Ap_Betweeness_natural=pd.DataFrame([dict(record) for record in session.run(query)])
    Ap_Betweeness_natural=Ap_Betweeness_natural[['cik','Ap_Betweeness_natural']]
Ap_Betweeness_natural

query='''CALL gds.betweenness.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'REVERSE',
      properties: {}
    }
  },
  samplingSize: 100
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name,  score
RETURN cik,  score as Ap_Betweeness_reverse
ORDER BY Ap_Betweeness_reverse DESC
'''

with driver.session(database="graphoptions") as session:
    Ap_Betweeness_reverse=pd.DataFrame([dict(record) for record in session.run(query)])
    Ap_Betweeness_reverse=Ap_Betweeness_reverse[['cik','Ap_Betweeness_reverse']]
Ap_Betweeness_reverse

## Pagerank 
Measures the transitive influence or connectivity of nodes

query='''CALL gds.pageRank.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'NATURAL',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value',
  dampingFactor: 0.85,
  maxIterations: 20
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name,  score
RETURN cik,  score as Page_Rank_wt_natural
ORDER BY Page_Rank_wt_natural DESC
'''

with driver.session(database="graphoptions") as session:
    Page_Rank_wt_natural=pd.DataFrame([dict(record) for record in session.run(query)])
    Page_Rank_wt_natural=Page_Rank_wt_natural[['cik','Page_Rank_wt_natural']]
Page_Rank_wt_natural

query='''CALL gds.pageRank.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'REVERSE',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value',
  dampingFactor: 0.85,
  maxIterations: 20
})) YIELD nodeId, score
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name,  score
RETURN cik,  score as Page_Rank_wt_reverse
ORDER BY Page_Rank_wt_reverse DESC
'''

with driver.session(database="graphoptions") as session:
    Page_Rank_wt_reverse=pd.DataFrame([dict(record) for record in session.run(query)])
    Page_Rank_wt_reverse=Page_Rank_wt_reverse[['cik','Page_Rank_wt_reverse']]
Page_Rank_wt_reverse

## HITS
Link analysis algorithm that rates nodes based on two scores, a hub score and an authority score.

query='''CALL gds.alpha.hits.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'REVERSE',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value',
  hitsIterations: 20
})) YIELD nodeId, values
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name, values.auth AS authScore, values.hub AS hubScore
RETURN cik, authScore as authScore_r, hubScore as hubScore_r
ORDER BY authScore DESC'''
with driver.session(database="graphoptions") as session:
    HITS_wt_reverse=pd.DataFrame([dict(record) for record in session.run(query)])
    HITS_wt_reverse=HITS_wt_reverse[['cik','authScore_r','hubScore_r']]
HITS_wt_reverse

query='''CALL gds.alpha.hits.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'NATURAL',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value',
  hitsIterations: 20
})) YIELD nodeId, values
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name, values.auth AS authScore, values.hub AS hubScore
RETURN cik, authScore as authScore_n, hubScore as hubScore_n
ORDER BY authScore DESC'''
with driver.session(database="graphoptions") as session:
    HITS_wt_natural=pd.DataFrame([dict(record) for record in session.run(query)])
    HITS_wt_natural=HITS_wt_natural[['cik','authScore_n','hubScore_n']]
HITS_wt_natural

## Closeness, it is unweighted
Detect nodes that are able to spread information very efficiently through a graph

query='''CALL gds.alpha.closeness.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'NATURAL',
      properties: {}
    }
  }
})) YIELD nodeId, centrality AS score
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name, score
RETURN cik,score as Closeness_natural
ORDER BY Closeness_natural DESC
'''
with driver.session(database="graphoptions") as session:
    Closeness_natural=pd.DataFrame([dict(record) for record in session.run(query)])
    Closeness_natural=Closeness_natural[['cik','Closeness_natural']]
Closeness_natural

query='''CALL gds.alpha.closeness.stream(({
  nodeProjection: '*',
  relationshipProjection: {
    relType: {
      type: '*',
      orientation: 'REVERSE',
      properties: {}
    }
  }
})) YIELD nodeId, centrality AS score
WITH gds.util.asNode(nodeId).id AS cik, gds.util.asNode(nodeId).company AS name, score
RETURN cik, score as Closeness_reverse
ORDER BY Closeness_reverse DESC
'''
with driver.session(database="graphoptions") as session:
    Closeness_reverse=pd.DataFrame([dict(record) for record in session.run(query)])
    Closeness_reverse=Closeness_reverse[['cik','Closeness_reverse']]
Closeness_reverse

## Degree centrality for call options using projected graph

query='''call gds.graph.drop('put')'''
with driver.session(database="graphoptions") as session:
    session.run(query)

query='''call gds.graph.drop('call')'''
with driver.session(database="graphoptions") as session:
    session.run(query)

query='''CALL gds.graph.create.cypher(
  'put',
  ' MATCH (n) RETURN toInteger(properties(n).id) AS id, labels(n) AS labels',
  ' MATCH (n)-[r:Holds]->(m)
    WHERE  r.putCall="Put"
    RETURN toInteger(properties(n).id) as source, toInteger(properties(m).id) as target,type(r) AS type, r.value as putvalue
'
)
YIELD
  graphName AS graph, nodeCount AS nodes, relationshipCount AS rels'''
with driver.session(database="graphoptions") as session:
    session.run(query)


query='''CALL gds.graph.create.cypher(
  'call',
  ' MATCH (n) RETURN toInteger(properties(n).id) AS id, labels(n) AS labels',
  ' MATCH (n)-[r:Holds]->(m)
    WHERE  r.putCall="Call"
    RETURN toInteger(properties(n).id) as source, toInteger(properties(m).id) as target,type(r) AS type, r.value as callvalue
'
)
YIELD
  graphName AS graph, nodeCount AS nodes, relationshipCount AS rels'''
with driver.session(database="graphoptions") as session:
    session.run(query)

query="""CALL gds.degree.stream('call',(
{orientation: 'NATURAL'
,  relationshipWeightProperty: 'callvalue'
})) YIELD nodeId, score
WITH nodeId AS cik,  score as Degree_natural_wt_call
RETURN cik,  Degree_natural_wt_call
ORDER BY Degree_natural_wt_call DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_natural_wt_call=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_natural_wt_call=Degree_natural_wt_call[['cik','Degree_natural_wt_call']]
Degree_natural_wt_call['cik']=Degree_natural_wt_call['cik'].apply(str)
Degree_natural_wt_call

query="""CALL gds.degree.stream('call',(
{orientation: 'REVERSE'
,  relationshipWeightProperty: 'callvalue'
})) YIELD nodeId, score
WITH nodeId AS cik,  score as Degree_reverse_wt_call
RETURN cik,  Degree_reverse_wt_call
ORDER BY Degree_reverse_wt_call DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_reverse_wt_call=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_reverse_wt_call=Degree_reverse_wt_call[['cik','Degree_reverse_wt_call']]
Degree_reverse_wt_call['cik']=Degree_reverse_wt_call['cik'].apply(str)
Degree_reverse_wt_call

## Degree centrality for put options using projected graph

query="""CALL gds.degree.stream('put',(
{orientation: 'NATURAL'
,  relationshipWeightProperty: 'putvalue'
})) YIELD nodeId, score
WITH nodeId AS cik,  score as Degree_natural_wt_put
RETURN cik,  Degree_natural_wt_put
ORDER BY Degree_natural_wt_put DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_natural_wt_put=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_natural_wt_put=Degree_natural_wt_put[['cik','Degree_natural_wt_put']]
Degree_natural_wt_put['cik']=Degree_natural_wt_put['cik'].apply(str)
Degree_natural_wt_put

query="""CALL gds.degree.stream('put',(
{orientation: 'REVERSE'
,  relationshipWeightProperty: 'putvalue'
})) YIELD nodeId, score
WITH nodeId AS cik,  score as Degree_reverse_wt_put
RETURN cik,  Degree_reverse_wt_put
ORDER BY Degree_reverse_wt_put DESC
"""

with driver.session(database="graphoptions") as session:
    Degree_reverse_wt_put=pd.DataFrame([dict(record) for record in session.run(query)])
    Degree_reverse_wt_put=Degree_reverse_wt_put[['cik','Degree_reverse_wt_put']]
Degree_reverse_wt_put['cik']=Degree_reverse_wt_put['cik'].apply(str)
Degree_reverse_wt_put

## Joining all the tables

dfs = [Degree_natural, Degree_reverse, Degree_natural_wt, Degree_reverse_wt, Betweeness, Betweeness_reverse, Ap_Betweeness_natural, Ap_Betweeness_reverse, Page_Rank_wt_natural, Page_Rank_wt_reverse, HITS_wt_reverse, HITS_wt_natural, Closeness_natural, Closeness_reverse,Degree_natural_wt_call,Degree_reverse_wt_call,Degree_natural_wt_put,Degree_reverse_wt_put]
from functools import reduce
df_final = reduce(lambda left,right: pd.merge(left,right,on='cik'), dfs)
df_final.to_csv(path+ 'df_final_graphoptions.csv', index=False)

df_final.head()
