==> Master_Thesis/Step 5.1 Community_detection_graph_wo_options.ipynb <==
from neo4j import GraphDatabase
import pandas as pd
from tabulate import tabulate
import os
import matplotlib
matplotlib.use('TkAgg')
import matplotlib.pyplot as plt

directory = os.getcwd()
path = directory + '/Data/'

## Connecting to database in neo4j

driver=GraphDatabase.driver("bolt://localhost:7687",auth=("neo4j",""))
session = driver.session(database="graphwooptions")

## Louvain 

query='''CALL gds.louvain.write({
  nodeProjection: 'company',
  relationshipProjection: {
    relType: {
      type: 'Holds',
      orientation: 'REVERSE',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value',
  includeIntermediateCommunities: false,
  seedProperty: '',
  writeProperty: 'louvain'
});'''
with driver.session(database="graphwooptions") as session:
    session.run(query)

## Label propagation algorithm

query='''CALL gds.labelPropagation.write({
  nodeProjection: 'company',
  relationshipProjection: {
    relType: {
      type: 'Holds',
      orientation: 'REVERSE',
      properties: {
        value: {
          property: 'value',
          defaultValue: 1
        }
      }
    }
  },
  relationshipWeightProperty: 'value',
  seedProperty: '',
  writeProperty: 'lpa'
});'''
with driver.session(database="graphwooptions") as session:
    session.run(query)

## Strongly connected components

query='''CALL gds.alpha.scc.write({
  nodeProjection: 'company',
  relationshipProjection: {
    relType: {
      type: 'Holds',
      orientation: 'REVERSE',
      properties: {}
    }
  },
  writeProperty: 'scc'
});'''
with driver.session(database="graphwooptions") as session:
    session.run(query)

## Analysing scc, the biggest community of scc has 129 nodes and has scc=14082

query='''MATCH (n)-[rel:Holds]->(m) where n.scc=14082 RETURN distinct n.id as id, n.company as company, n.scc as scc'''
    
    
with driver.session(database="graphwooptions") as session:
    scc=pd.DataFrame([dict(record) for record in session.run(query)])
    scc=scc[['scc','company','id']]

scc.to_csv(path+'scc.csv')
