==> Master_Thesis/Step 6. Downsizing .csv to .pkl.ipynb <==
import os
import pandas as pd

directory = os.getcwd()
path = directory + '/Data/'

## Reading files

header_list = ["cusip",'cik','titleOfClass','nameOfIssuer','putCall','investmentDiscretion','value','shrsOrPrnAmt.sshPrnamt',
               'shrsOrPrnAmt.sshPrnamtType','votingAuthority.Sole','votingAuthority.Shared','votingAuthority.None','id','Form']
cik_header=["nameOfIssuer","CIK"]

Filing_final=pd.read_csv(path + 'Filing_final.csv')
Hold_final=pd.read_csv(path + 'Hold_final.csv',names=header_list,low_memory=False, header=0)
Filings_df=pd.read_csv(path + 'Filings_df.csv')
Holdings=pd.read_csv(path + 'Holdings.csv')
Holdings_df=pd.read_csv(path + 'Holdings_df.csv')
cik_v3=pd.read_csv(path+'cik_v3.csv',low_memory=False)
Relationships=pd.read_csv(path + 'Relationships.csv')
Relationships_wo_options=pd.read_csv(path + 'Relationships_wo_options.csv')

## Converting to .pkl files to reduce size

Filing_final.to_pickle(path + "Filing_final.pkl", compression='bz2')
Hold_final.to_pickle(path + "Hold_final.pkl", compression='bz2')
Filings_df.to_pickle(path + "Filings_df.pkl", compression='bz2')
Holdings.to_pickle(path + "Holdings.pkl", compression='bz2')
Holdings_df.to_pickle(path + "Holdings_df.pkl", compression='bz2')
cik_v3.to_pickle(path + "cik_v3.pkl", compression='bz2')
Relationships.to_pickle(path + "Relationships.pkl", compression='bz2')
Relationships_wo_options.to_pickle(path + "Relationships_wo_options.pkl", compression='bz2')
